refactor(ast): simplify declaration conflict reporting
Store append-time declaration errors directly on Code and remove the redundant conflict translation layer and internal-only regression. Existing compiler tests retain coverage for duplicate functions, overloads, and the shared constant namespace.
